install script

+++ nvim-install.sh
@@ -1,16 +1,13 @@
 #!/bin/sh
 
-echo "Running this script will erase your current neovim configuration, are you sure you want to continue? [Y/n]: "
+read -p "Running this script will erase your current neovim configuration, are you sure you want to continue? [Y/n]: " prompt
 
-read -r response
+prompt="${prompt:-Y}"
+prompt=$(echo "$prompt" | tr '[:upper:]' '[:lower:]')
 
-case "$response" in
-    [nN]|[nN][oO])  
-        echo "Aborting."
-        exit 1
-        ;;
-    *) 
-        if [ -d "$HOME/.config/nvim" ]; then
+if [ "$prompt" = "y" ]; then
+
+	if [ -d "$HOME/.config/nvim" ]; then
             rm -rf "$HOME/.config/nvim"
         fi
 
@@ -18,9 +15,14 @@ case "$response" in
         mv nvim-conf nvim
         mv nvim ~/.config
 	rm ~/.config/nvim/autoload/plug.vim	# removes backup
-	wget https://raw.githubusercontent.com/junegunn/vim-plug/refs/heads/master/plug.vim -P ~/.config/nvim/autoload	#gets the latest version of vim-plug
+	wget https://raw.githubusercontent.com/junegunn/vim-plug/refs/heads/master/plug.vim -P ~/.config/nvim/autoload	# gets the latest version of vim-plug
         rm ~/.config/nvim/nvim-install.sh	# removes redundant install script
-	;;
-esac
+
+else
+
+echo "aborting."
+exit 1
+
+fi
 
 exit 0
